+++ modules/template/widgets/views/templateElementAdminRow.php
@@ -1,14 +1,24 @@
 <?php
 
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ */
+
+use humhub\modules\custom_pages\modules\template\models\TemplateElement;
 use humhub\widgets\bootstrap\Badge;
 use humhub\widgets\bootstrap\Button;
 use yii\helpers\Html;
 
-/* @var $model humhub\modules\custom_pages\modules\template\models\TemplateElement */
+/* @var $model TemplateElement */
 ?>
-<tr data-template-element-definition="<?= $model->id ?>" >
+<tr data-template-element-definition="<?= $model->id ?>">
+    <td>
+        <?= Html::encode($model->title) ?>
+    </td>
     <td class="text-nowrap">
-        #<strong><?= Html::encode($model->name) ?> </strong>
+        {{ <?= Html::encode($model->name) ?> }}
     </td>
     <td>
         <?= Badge::success($model->getLabel()) ?>
@@ -18,11 +28,8 @@ use yii\helpers\Html;
             <?= Badge::success(Yii::t('CustomPagesModule.base', 'Default')) ?>
         <?php endif; ?>
     </td>
-
     <td>
     <?php if ($model->template->canEdit()) : ?>
-        <?= Button::primary()->icon('pencil')->sm()
-            ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
         <?= Button::danger()->icon('times')->sm()
             ->action('deleteElementSubmit', ['/custom_pages/template/admin/delete-element', 'id' => $model->id])
             ->confirm(
@@ -30,6 +37,8 @@ use yii\helpers\Html;
                 Yii::t('CustomPagesModule.template', 'Do you really want to delete this element? <br />The deletion will affect all pages using this template.'),
                 Yii::t('CustomPagesModule.base', 'Delete'),
             ) ?>
+        <?= Button::primary()->icon('pencil')->sm()
+            ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
     <?php else : ?>
         <?= Button::accent()->icon('eye')->sm()
             ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>