+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 
 1.11.0-beta.12 (Unreleased)
 --------------------------------
+- Enh #433: Change template overview
 - Enh #402: HTML pages not allowed in Spaces
 - Enh #457: Fix adding items to empty container
 - Fix #464: Add safe migrations